# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ecommerce REST API built using Go Lang, PSQL, REST API following Clean Architecture. Hosted with Ngnix, AWS EC2 and RDS",
        "title": "Ecommerce REST API",
        "contact": {
            "name": "Nadeem Fahad",
            "url": "https://github.com/Nadeem1815",
            "email": "nadeemf408@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/admin/block_user/:user_id": {
            "patch": {
                "description": "Admin Blocked for user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Admin Blocked USer",
                "operationId": "admin-block-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user to be blocked",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
...
        "/admin/unblock_user/:user_id": {
            "patch": {
                "description": "Admin UnBlocked for user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Admin UnBlocked USer",
                "operationId": "admin-unblock-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of the user to be Unblocked",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
...
        "/user/address": {
            "post": {
                "description": "User add address field",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Add Address",
                "operationId": "user-addaddress",
                "parameters": [
                    {
                        "description": "User add Address",
                        "name": "user_address",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_model.AddressInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
...
        "/user/home": {
            "get": {
                "description": "when login user can see home page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "HomePage",
                "operationId": "home",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
        "/user/login/email": {
            "post": {
                "description": "User login With Email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "LoginWithEmail",
                "operationId": "login-with-email",
                "parameters": [
                    {
                        "description": "User login credentials",
                        "name": "User_details",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_model.UserLoginEmail"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "description": "Logs out a logged-in user from the E-commerce web api user panel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User Logout",
                "operationId": "user-logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "New User  can Registration.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User SignUp",
                "operationId": "user-signup",
                "parameters": [
                    {
                        "description": "Register User",
                        "name": "user_details",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_model.UsarDataInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/github_com_nadeem1815_premium-watch_pkg_utils_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_nadeem1815_premium-watch_pkg_utils_model.AddressInput": {
            "type": "object",
            "properties": {
                "district": {
                    "type": "string"
                },
                "house_name": {
                    "type": "string"
                },
                "landmark": {
                    "type": "string"
                },
                "pincode": {
                    "type": "integer"
                },
                "state": {
                    "type": "string"
                },
                "street": {
                    "type": "string"
                }
            }
        },
...
        "github_com_nadeem1815_premium-watch_pkg_utils_model.UsarDataInput": {
            "type": "object",
            "required": [
                "email_id",
                "password",
                "phone"
            ],
            "properties": {
                "email_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8
                },
                "phone": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
        "github_com_nadeem1815_premium-watch_pkg_utils_model.UserLoginEmail": {
            "type": "object",
            "required": [
                "email_id",
                "password"
            ],
            "properties": {
                "email_id": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 64,
                    "minLength": 8
                }
            }
        },
        "github_com_nadeem1815_premium-watch_pkg_utils_response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}